+++ packages/faceted-search/stories/facetedSearch.story.js
@@ -5,73 +5,20 @@ import IconsProvider from '@talend/react-components/lib/IconsProvider';
 import FacetedSearch from '../src';
 import { FacetedSearchIcon } from '../src/components/FacetedSearchIcon';
 
-const badgeName = {
-	properties: {
-		attribute: 'name',
-		initialOperatorOpened: true,
-		initialValueOpened: false,
-		label: 'Name',
-		operator: {},
-		operators: [],
-		type: 'text',
-	},
-	metadata: {
-		badgePerFacet: 'N',
-		entitiesPerBadge: '1',
-		operators: ['contains', 'equals', 'notEquals'],
-	},
-};
-
-const badgeConnectionType = {
-	properties: {
-		attribute: 'connection.type',
-		initialOperatorOpened: true,
-		initialValueOpened: false,
-		label: 'Connection type',
-		operator: {},
-		operators: [],
-		type: 'checkbox',
-	},
-	metadata: {
-		badgePerFacet: '1',
-		entitiesPerBadge: 'N',
-		values: [
-			{ id: 'amazon_s3', label: 'Amazon S3' },
-			{ id: 'hdfs', label: 'HDFS' },
-			{ id: 'kafka', label: 'Kafka' },
-			{ id: 'localcon', label: 'Local connection' },
-			{ id: 'salesforce', label: 'Salesforce' },
-			{ id: 'aws_kinesis', label: 'AWS kinesis' },
-		],
-		operators: ['in'],
-	},
-};
-
-const badgePrice = {
-	properties: {
-		attribute: 'price',
-		initialOperatorOpened: true,
-		initialValueOpened: false,
-		label: 'Price',
-		operator: {},
-		operators: [],
-		type: 'number',
-	},
-	metadata: {
-		badgePerFacet: 'N',
-		entitiesPerBadge: '1',
-		operators: [
-			'equals',
-			'notEquals',
-			'greaterThan',
-			'greaterThanOrEquals',
-			'lessThan',
-			'lessThanOrEquals',
-		],
-	},
-};
+import {
+	badgeConnectionType,
+	badgeName,
+	badgePrice,
+	badgeWithVeryLongName,
+} from './badgesDefinitions.story';
 
 const badgesDefinitions = [badgeName, badgeConnectionType, badgePrice];
+const lotsOfBadgesDefinitions = [];
+let i = 0;
+while (i < 50) {
+	lotsOfBadgesDefinitions.push(badgeName);
+	i += 1;
+}
 
 const paddingLeft = { paddingLeft: '10px' };
 
@@ -113,6 +60,28 @@ storiesOf('FacetedSearch', module)
 						/>
 					))
 				}
+			</FacetedSearch.Faceted>
+		</div>
+	))
+	.add('basic search with lots of badges definitions', () => (
+		<div style={{ height: '5.5rem' }}>
+			<IconsProvider />
+			<FacetedSearch.Faceted id="my-faceted-search">
+				<FacetedSearch.BasicSearch
+					badgesDefinitions={lotsOfBadgesDefinitions}
+					onSubmit={action('onSubmit')}
+				/>
+			</FacetedSearch.Faceted>
+		</div>
+	))
+	.add('basic search with badge with very long name', () => (
+		<div style={{ height: '5.5rem' }}>
+			<IconsProvider />
+			<FacetedSearch.Faceted id="my-faceted-search">
+				<FacetedSearch.BasicSearch
+					badgesDefinitions={[badgeWithVeryLongName, badgeConnectionType, badgeName, badgePrice]}
+					onSubmit={action('onSubmit')}
+				/>
 			</FacetedSearch.Faceted>
 		</div>
 	));